fix: brighter target text + full date format (July 16, 2026) throughout - Target text changed from text-xs text-muted-foreground to text-sm text-foreground/70 - Model updated date uses full format: July 12, 2026 - FPR/FNR trend: July 10, July 11 (full month name) - Defect trends: July 10, July 11 (full month name) - Recent decisions: July 16, 2026, 10:45 AM

+++ frontend/src/components/dashboard/DefectDashboard.tsx
@@ -68,7 +68,7 @@ export function DefectDashboard() {
   }
   const trendData = Object.entries(trendByDate).map(([date, classes]) => {
     const d = new Date(date + "T00:00:00");
-    const formatted = d.toLocaleDateString("en-US", { month: "short", day: "numeric" });
+    const formatted = d.toLocaleDateString("en-US", { month: "long", day: "numeric" });
     return { date: formatted, ...classes };
   });
   const trendClasses = [...new Set(data.trends.map((t) => t.defect_class))];
